requirements.txt
torch==2.5.1+cu118  # PyTorch for deep learning
torchvision==0.20.1  # PyTorch vision library
hydra-core==1.3.2  # Configuration management for SAM2
mediapipe==0.10.20  # For wrist detection
opencv-python==4.11.0.86  # For video and image processing
Pillow==10.4.0  # Image manipulation
ffmpeg-python==0.2.0  # Python bindings for FFmpeg
SAM-2 @ git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4  # SAM2 segmentation model